requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.32.*
sqlmodel==0.0.22
pydantic==2.9.*
pydantic-settings==2.6.*
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
slowapi==0.1.9

# Tools
ruff==0.7.*
mypy==1.13.*
bandit==1.7.*
pytest==8.3.*
httpx==0.27.*
